+++ test/badges.test.ts
@@ -20,7 +20,7 @@ const specUri2 = 'another spec uri'
 const reasonChoice = 1
 const emptyBytes32String = ethers.utils.formatBytes32String('')
 const errNotOwner = 'Ownable: caller is not the owner'
-const errSpecNotRegistered = 'mint: spec is not registered'
+const errSpecNotRegistered = 'spec is not registered'
 const errSpecAlreadyRegistered = 'createSpec: spec already registered'
 const errNotRaftOwner = 'createSpec: unauthorized'
 const errInvalidSig = 'safeCheckAgreement: invalid signature'
@@ -30,6 +30,11 @@ const errNotRevoked = 'reinstateBadge: badge not revoked'
 const errBadgeAlreadyRevoked = 'revokeBadge: badge already revoked'
 const errMerkleAlreadyUsed = 'newSafeCheckAgreement: already used'
 
+const errNoSpecUris = 'refreshMetadata: no spec uris provided'
+const errUnauthorizedGive = 'give: unauthorized'
+const errUnauthorizedTake = 'take: unauthorized issuer'
+const errCannotGiveToSelf = 'give: cannot give to self'
+const err721InvalidTokenId = 'ERC721: invalid token ID'
 let deployed: any
 
 // fix ts badgesProxy: any
@@ -40,7 +45,7 @@ async function createSpec(badgesProxy: any, specUri: string, raftTokenId: BigNum
   return await getSpecCreatedEventLogData(txn.hash, badgesProxy)
 }
 
-async function mintBadge() {
+async function mintBadgeWithTake() {
   // deploy contracts
   const { badgesProxy, raftProxy, typedData, issuer, claimant, owner } = deployed
   const specUri = typedData.value.tokenURI
@@ -75,6 +80,31 @@ async function mintBadge() {
   const uriOfToken = await badgesProxy.tokenURI(transferEventData.tokenId)
   expect(uriOfToken).to.equal(typedData.value.tokenURI)
   return { raftTokenId, badgeId: transferEventData.tokenId }
+}
+
+async function mintBadgeWithGive(walletCallingGive: SignerWithAddress) {
+  const { badgesProxy, typedData, issuer, claimant } = deployed
+  // flip the values of active and passive for "give" functionality. Default values are for "take"
+  typedData.value.active = issuer.address
+  typedData.value.passive = claimant.address
+
+  // for "give" the claimant creates the signature
+  const { compact } = await getSignature(typedData.domain, typedData.types, typedData.value, claimant)
+
+  // and the issuer calls give
+  const txn = await badgesProxy
+    .connect(walletCallingGive)
+    .give(typedData.value.passive, typedData.value.tokenURI, compact)
+  await txn.wait()
+
+  const transferEventData = await getTransferEventLogData(txn.hash, badgesProxy)
+
+  expect(transferEventData.to).equal(typedData.value.passive)
+  expect(transferEventData.tokenId).gt(0)
+
+  expect(await badgesProxy.ownerOf(transferEventData.tokenId)).to.equal(claimant.address)
+  expect(await badgesProxy.balanceOf(claimant.address)).to.equal(1)
+  expect(await badgesProxy.tokenURI(transferEventData.tokenId)).to.equal(typedData.value.tokenURI)
 }
 
 async function mintRaftToken(raftProxy: any, toAddress: string, raftTokenUri: string, signer: SignerWithAddress) {
@@ -377,61 +407,6 @@ describe('Merkle minting', () => {
   })
 })
 
-describe('Proxy upgrades', () => {
-  it('Should upgrade the Raft contract then create raft/spec/badge', async () => {
-    // deploy contracts
-    const { raftProxy } = deployed
-
-    const raftV2 = await ethers.getContractFactory('RaftV2')
-    const upgradedV2Contract = await upgrades.upgradeProxy(raftProxy.address, raftV2)
-    await upgradedV2Contract.deployed()
-
-    const v2 = await upgradedV2Contract.getVersion()
-    // expect(v2).equal(2)
-    await mintBadge()
-  })
-
-  it('Should upgrade the Badge contract then create raft/spec/badge', async () => {
-    // deploy contracts
-    const { badgesProxy } = deployed
-
-    const badgesV2 = await ethers.getContractFactory('BadgesV2')
-    const upgradedV2Contract = await upgrades.upgradeProxy(badgesProxy.address, badgesV2)
-    await upgradedV2Contract.deployed()
-
-    const v2 = await upgradedV2Contract.getVersion()
-    expect(v2).equal(2)
-
-    await mintBadge()
-  })
-
-  it('Should upgrade then instantiate new variable right after', async () => {
-    const { badgesProxy } = deployed
-
-    const badgesV2 = await ethers.getContractFactory('BadgesV2')
-    const upgradedV2Contract = await upgrades.upgradeProxy(badgesProxy.address, badgesV2)
-    await upgradedV2Contract.deployed()
-    await upgradedV2Contract.setNewVar()
-    const newVar = await upgradedV2Contract.myNewVar()
-    expect(newVar).equal(9)
-    await expect(upgradedV2Contract.setNewVar()).to.be.revertedWith('Var is already set')
-  })
-
-  it('Should upgrade the SpecDataHolder contract then create raft/spec/badge', async () => {
-    // deploy contracts
-    const { specDataHolderProxy } = deployed
-
-    const specDataHolderV2 = await ethers.getContractFactory('SpecDataHolderV2')
-    const upgradedV2Contract = await upgrades.upgradeProxy(specDataHolderProxy.address, specDataHolderV2)
-    await upgradedV2Contract.deployed()
-
-    const v2 = await upgradedV2Contract.getVersion()
-    expect(v2).equal(2)
-
-    await mintBadge()
-  })
-})
-
 describe('Badge Specs', () => {
   it('should register a spec successfully', async function () {
     // deploy contracts
@@ -536,7 +511,7 @@ describe('Badges', async function () {
   })
 
   it('should successfully mint badge', async function () {
-    await mintBadge()
+    await mintBadgeWithTake()
   })
 
   it('should fail when trying to claim using a voucher from another issuer for the same spec', async function () {
@@ -645,12 +620,12 @@ describe('Badges', async function () {
     // get signature
     const { compact } = await getSignature(typedData.domain, typedData.types, typedData.value, issuer)
     await expect(badgesProxy.connect(claimant).take(typedData.value.passive, specUri, compact)).to.be.revertedWith(
-      errSpecNotRegistered
+      err721InvalidTokenId
     )
   })
 
   it('Should revoke a badge, confirm its revoked, then reinstate', async () => {
-    const { raftTokenId, badgeId } = await mintBadge()
+    const { raftTokenId, badgeId } = await mintBadgeWithTake()
     const { badgesProxy, issuer } = deployed
     // revoke the badge
     await badgesProxy.connect(issuer).revokeBadge(raftTokenId, badgeId, reasonChoice)
@@ -666,7 +641,7 @@ describe('Badges', async function () {
   })
 
   it('Should fail to revoke a badge if passed an invalid tokenId', async () => {
-    const { raftTokenId, badgeId } = await mintBadge()
+    const { raftTokenId, badgeId } = await mintBadgeWithTake()
     const { badgesProxy, issuer } = deployed
 
     await expect(badgesProxy.connect(issuer).revokeBadge(raftTokenId, 123, reasonChoice)).to.be.revertedWith(
@@ -675,7 +650,7 @@ describe('Badges', async function () {
   })
 
   it('Should fail to reinstate a badge that is not revoked', async () => {
-    const { raftTokenId, badgeId } = await mintBadge()
+    const { raftTokenId, badgeId } = await mintBadgeWithTake()
     const { badgesProxy, issuer } = deployed
 
     expect(await badgesProxy.connect(issuer).isBadgeValid(badgeId)).to.equal(true)
@@ -684,7 +659,7 @@ describe('Badges', async function () {
   })
 
   it('Should fail to revoke a badge if its already revoked', async () => {
-    const { raftTokenId, badgeId } = await mintBadge()
+    const { raftTokenId, badgeId } = await mintBadgeWithTake()
     const { badgesProxy, issuer } = deployed
     // revoke the badge
     await badgesProxy.connect(issuer).revokeBadge(raftTokenId, badgeId, reasonChoice)
@@ -694,5 +669,58 @@ describe('Badges', async function () {
     await expect(badgesProxy.connect(issuer).revokeBadge(raftTokenId, badgeId, reasonChoice)).to.be.revertedWith(
       errBadgeAlreadyRevoked
     )
+  })
+
+  it('Should not refreshMetadata if caller isnt the owner', async () => {
+    const { badgesProxy, issuer } = deployed
+    await expect(badgesProxy.connect(issuer).refreshMetadata([specUri])).to.be.revertedWith(errNotOwner)
+  })
+
+  it('Should not refreshMetadata no spec uris are passed in to be refreshed', async () => {
+    const { badgesProxy, owner } = deployed
+    const { raftTokenId, badgeId } = await mintBadgeWithTake()
+    await expect(badgesProxy.connect(owner).refreshMetadata([])).to.be.revertedWith(errNoSpecUris)
+  })
+
+  it('Should call give with the happy path', async () => {
+    const { badgesProxy, raftProxy, typedData, issuer, claimant, owner } = deployed
+    const specUri = typedData.value.tokenURI
+
+    const { raftTokenId } = await mintRaftToken(raftProxy, issuer.address, specUri, owner)
+
+    await createSpec(badgesProxy, specUri, raftTokenId, issuer)
+
+    await mintBadgeWithGive(issuer)
+  })
+
+  it('Should require that give is called by raft owner', async () => {
+    const { badgesProxy, raftProxy, typedData, issuer, randomSigner, owner } = deployed
+    const specUri = typedData.value.tokenURI
+
+    const { raftTokenId } = await mintRaftToken(raftProxy, issuer.address, specUri, owner)
+
+    await createSpec(badgesProxy, specUri, raftTokenId, issuer)
+
+    // the person calling "give" here is a random signer, not the owner of the raft token
+    await expect(mintBadgeWithGive(randomSigner)).to.be.revertedWith(errInvalidSig)
+  })
+
+  it('Should shouldnt allow someone to call give to themself', async () => {
+    const { badgesProxy, raftProxy, typedData, issuer, claimant, owner } = deployed
+    const specUri = typedData.value.tokenURI
+
+    const { raftTokenId } = await mintRaftToken(raftProxy, issuer.address, specUri, owner)
+
+    await createSpec(badgesProxy, specUri, raftTokenId, issuer)
+
+    typedData.value.active = issuer.address
+    typedData.value.passive = claimant.address
+
+    const { compact } = await getSignature(typedData.domain, typedData.types, typedData.value, claimant)
+
+    // first param of give here is the "to" address (which is the same as the issuer)
+    await expect(
+      badgesProxy.connect(issuer).give(typedData.value.active, typedData.value.tokenURI, compact)
+    ).to.be.revertedWith(errCannotGiveToSelf)
   })
 })