+++ changelog.md
@@ -2,6 +2,18 @@
 
 We're using https://semver.org/
 
+## 2.4.0
+
+- Update the "give" function to make sure that the caller is the owner of the Raft token associated with the targeted badge spec
+
+## 2.3.0
+
+- Add a function that enables us refresh a badge's metadata by emitting an event
+
+## 2.2.0
+
+- Add revocation and reinstatement of badges
+
 ## 2.1.0
 
 - Added MetadataUpdate event in Raft.sol when setTokenURI is called